KDA/GDN cleanup
Start a running cleanup PR for linear-attention maintainability and documentation.
This first pass changes documentation and test placement, not production behavior.

- Add navigation and focused context-parallel sections to the linear-attention reference.
- Clarify fixed storage versus changing routing under CUDA Graph replay, including why
  captured host terminal indices do not follow new document layouts.
- Move the existing summary composition/identity test into the CPU routing suite so
  it runs without CuTeDSL installed; preserve its assertions and the separate CUDA
  precision-mode regression.

Validation:
- pytest -n 6 test/test_context_parallel_plan.py: 92 passed on CPU without Cutlass.
- Applicable prek hooks, ruff check/format, and git diff --check passed.
- New navigation links checked; production executable AST unchanged.
- No GPU tests or full documentation build run for this documentation/test-placement pass.

Keep this PR in draft while additional focused cleanups are developed and validated.

stack-info: PR: https://github.com/meta-pytorch/attention-gym/pull/488, branch: drisspg/stack/132
